ver 6.0 Added command-object: GOTO.
Added search for command tracks with the status of future commands or the status of current command in the Thread Task class.
If there is no command with the current status, then the search for a suitable command for this role is started according to the timing and status of future commands.
If such a command object exists, then we receive some fields of this command and transfer the command to the current status.

Added the Invoker class that implements the "command pattern".
Added concrete implementation for the "GOTO" type object command.

To do:
Changing the state of objects (trucks) after executing the current command of the "GOTO" type with assignment of the command status to archive.